Builds on top of LVN and adds other optimizations.
This CR builds on top of the LVN algorithm implemented in the previous
CR and adds the following optimizations -
1/ Copy propagation - This was enabled by modifying the `Value` lookup
operation. Details can be seen in code comments for
`lvn::Value2ValueNumber::get` method.

2/ Constant propagation - This was achieved by keeping a map of value
numbers that correspond to canonical variables with constant/literal
values.

3/ Constant folding - This was achieved by using the information stored
in the map, mentioned in (2) and determining whether certain "foldable"
`Value`s can indeed be folded.