Adds utilities to convert bril functions to SSA form
This commit builds on top of the commit where we added utilities to
generate dominance information and adds support for converting bril
functions to SSA form. The newly added SSA utilities work in general
other than for the testcase in `examples/test/ssa/if-orig.bril`. The
issue seems to be that I am not storing the right information in the
dominance tree. Currently the dominance tree is represented as a graph
and for each node in the graph it only stores edges for the immediately
dominated nodes. Instead, I think, we need to store the edges for both
strictly and immediately dominated nodes.

Next steps are to fix this bug and then add support for SSA to bril
conversion.