Fixes the remaining bug in `to_ssa` implementation.
This CR fixes the bug in the `to_ssa` implementation, that was mentioned
in the last commit. This bug was just causing the `ssa/if-orig.bril`
test case to fail, but I think the other tests were only passing because
the CFG representations of the bril functions in these tests were not
exposing a critical flaw in my `DominanceTree` type.

My `DominanceTree` type stores the immediate dominators for all nodes.
The issue was that the immediate dominator sets stored by the tree were
incorrect. For each node, I was only considering the nodes that were
this node's predecessor in the CFG as potential candidates for immediate
dominators.

A node immediately dominates only those nodes that it strictly
dominates, but those strictly dominated by it don't strictly
dominate. For instance, let's say we have the following nodes in our dominance tree -
```
0: {1, 2, 3}
1: {2, 3}
2: {}
3: {}
```
The node `0` immediately dominated only those nodes that it
strictly dominates, minus the ones that its strictly dominated
children don't strictly dominate:
```
 {1, 2, 3} - {{2, 3} U {} U {}}
```