Get `from_ssa` implementation working.
This CR adds a basic implementation for the `from_ssa` function and
with it, essentially, an implementation for a SSA roundtrip.

The implementation leaves a number of things to be desired -
1. The bril code after a SSA roundtrip has a lot more redundant copies
   that could potentially be optimized away.
2. Even though the `Phi` nodes are removed the variables are still left
   in SSA form, i.e., with the `.1` type suffixes. I'm not sure what's
   usually supposed to happen to them and this is something that I WILL
   NEED TO FIGURE OUT!

Another thing that I realized was never implemented was, making sure
that each CFG has a unique entry block (and modifying it otherwise).
This caused failures in the `while.bril` test case. I left this as it is
for the time being.