Fixes a bug in CFG construction and adds types for dominance analysis
This commit fixed a few bugs in CFG construction from a function due to
which a block's successors and predecessors were not being set
correctly. This in turn was affecting the dominance analysis results.

I've also added unit tests to verify both CFG predecessors and sucessors
as well as CFG dominators. The lesson learnt during this session was a
hunbling one:

NO ONE CARES WHO YOU ARE, DO NOT SKIP UNIT TESTS OR YOU WILL BE MISERABLE.

Steps for next session
-----------------------
1. Implement the function to create a dominator tree.